tests - fixes after refactor
